separate audit code & remove useless code
