+++ audits/audit.py
@@ -1,11 +1,10 @@
 # query enrolments for a given student number join that certain student
 from datetime import datetime
-import json
-from copy import deepcopy
 
 from datamodel.models import Enrolment, Course
 from StudentTrackingSystemApp.configfuncs import excel_in_dict as xls_confs
 from StudentTrackingSystemApp.configfuncs import get_all_cores
+from audits.audit_data import AuditData
 
 UNPOP = "UNPOPULATED"
 
@@ -15,57 +14,6 @@ def _find_if(compare, list):
         if compare(elem):
             return elem
     return None
-
-
-class AuditData:
-    empty_progress_data = {"courses": [], "credit_hours": 0}
-    empty_progress = {
-        "completed": deepcopy(empty_progress_data),
-        "in_progress": deepcopy(empty_progress_data),
-    }
-
-    def __init__(self):
-        self.data = {
-            "target_student": {},
-            "progress": {},
-        }
-
-    def add_course(self, status: str, course: Course):
-        progress = self.data["progress"]
-        type_progress = progress.get(
-            course.course_type, deepcopy(AuditData.empty_progress)
-        )
-        if status not in type_progress:
-            type_progress[status] = deepcopy(AuditData.empty_progress_data)
-        type_progress[status]["courses"].append(course.course_code.replace("*", ""))
-        type_progress[status]["credit_hours"] += course.credit_hours
-        self.data["progress"][course.course_type] = type_progress
-
-    def remove_course(self, status: str, course: Course):
-        progress = self.data["progress"]
-        type_progress = progress.get(
-            course.course_type, deepcopy(AuditData.empty_progress)
-        )
-        if status not in type_progress:
-            return
-        try:
-            type_progress[status]["courses"].remove(course.course_code.replace("*", ""))
-            type_progress[status]["credit_hours"] -= course.credit_hours
-            self.data["progress"][course.course_type] = type_progress
-        except ValueError:
-            pass  # course already removed from lits, that's ok
-
-    def __getitem__(self, key):
-        return self.data[key]
-
-    def __setitem__(self, key, value):
-        self.data[key] = value
-
-    def __str__(self):
-        return json.dumps(self.data)
-
-    def toJSON(self):
-        return json.dumps(self.data)
 
 
 def audit_student(student_number, mapped_courses=None):
@@ -121,35 +69,10 @@ def audit_student(student_number, mapped_courses=None):
     return (audit_response, mapped_courses)
 
 
-def _lookup_course_type(course):
-    replacements = xls_confs["replacements"].to_dict(orient="list")
-    course_types = xls_confs["valid-tags"].to_dict(orient="list")
-    exceptions = xls_confs["exceptions"].to_dict(orient="list")
-
-    # print("replacements", replacements)
-    # print("exceptions", exceptions)
-
-    course_type = None
-    joined_code = "".join(course.course_code.split("*"))
-
-    def _code_matches_course_in_type(target_course):
-
-        return joined_code.startswith(target_course)  # and not
-
-    # Match to valid tags
-    for type, courses_in_type in course_types.items():
-        if joined_code in exceptions[type]:
-            break
-
-        found = _find_if(_code_matches_course_in_type, courses_in_type)
-        if len(found):
-            return type
-            # iterate through found, checking if
-
-    return course_type
-
-
 def _target_student_data(student):
+    # years_in calculated as current year minus start year, plus
+    # the portion of the current year, in months, that have passed
+    # the current month is ignored, which is why now.month - 1 is used
     now = datetime.now()
     years_in = float(now.year - student.upload_set.upload_datetime.year)
     years_in = round(years_in + (now.month - 1) / 12, 1)
